# openapi.yaml
openapi: "3.1.0"
info:
  title: Agent Task Editor API
  version: "1.0.0"
servers:
  - url: http://localhost:8080/api/v1
components:
  schemas:
...
    AgentConfig:
      type: object
      properties:
        id:            { type: string }
        name:          { type: string }
        provider:      { type: string, enum: [claude_code, openai, anthropic, custom] }
        model:         { type: string }
        system_prompt: { type: string }
        labels:        { type: string, description: "JSON array of label names" }
        env:           { type: string, description: "JSON object of env vars" }
        max_tokens:    { type: integer }
        timeout_secs:  { type: integer }
        created_at:    { type: string, format: date-time }
        updated_at:    { type: string, format: date-time }
    Repo:
      type: object
      properties:
        id:          { type: string }
        name:        { type: string }
        path:        { type: string }
        remote_url:  { type: string, nullable: true }
        workflow_id: { type: string, nullable: true }
        created_at:  { type: string, format: date-time }
...
paths:
...
  /agents/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Get an agent config
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentConfig" }
    put:
      summary: Update an agent config
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentConfig" }
    delete:
      summary: Delete an agent config
      responses:
        "204": { description: Deleted }
...
  /repos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Get a repo
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Repo" }
    delete:
      summary: Remove a repo
      responses:
        "204": { description: Removed }
...